减少用户 marksCount，当取消标记时

+++ src/graphql/resolvers/user/unmark.ts
@@ -3,6 +3,7 @@ import { AppContext } from "../../../app/context";
 import { Connection } from "typeorm";
 import { PostUserMarkEntity } from "../../../entity/PostUserMarkEntity";
 import { getPayloadUserId } from "../../../functional/token/tokenservice";
+import { UserEntity } from "../../../entity/UserEntity";
 
 @InputType()
 export class UnmarkInput {
@@ -18,8 +19,13 @@ export async function unmark(
 
   const connection = container.resolve(Connection);
   const postUserMarkRepository = connection.getRepository(PostUserMarkEntity);
+  const userRepository = connection.getRepository(UserEntity);
   const userId = getPayloadUserId(tokenPayload);
 
-  await postUserMarkRepository.delete({ userId, postId });
+  const link = await postUserMarkRepository.findOne({ userId, postId });
+  if (link != null) {
+    await postUserMarkRepository.delete({ userId, postId });
+    await userRepository.decrement({ id: userId }, 'marksCount', 1);
+  }
   return 'success';
 }
